feat: add schedule components
