+++ components/shop/data.js
@@ -58,57 +58,3 @@ export const MM_PURCHASE_STEPS = [
     action: "Keep the proof of payment (transaction receipt) for verification.",
   },
 ];
-
-export const PYCON_DAY1_SCHEDULE = [
-  {
-    title: "Opening Keynotes",
-    duration: "09:00 - 09:30",
-    speaker: "John Doe",
-  },
-  {
-    title: "Break Out Sessions",
-    duration: "09:00 - 09:30",
-    speaker: "Moses Mark",
-  },
-  {
-    title: "What is Pycon",
-    duration: "09:00 - 09:30",
-    speaker: "James Karen",
-  },
-];
-
-export const PYCON_DAY2_SCHEDULE = [
-  {
-    title: "Opening Keynotes",
-    duration: "09:00 - 09:30",
-    speaker: "John Doe",
-  },
-  {
-    title: "Python IDEs",
-    duration: "09:00 - 09:30",
-    speaker: "Moses Mark",
-  },
-  {
-    title: "Python Overview",
-    duration: "09:00 - 09:30",
-    speaker: "James Karen",
-  },
-];
-
-export const PYCON_DAY3_SCHEDULE = [
-  {
-    title: "Opening Keynotes",
-    duration: "09:00 - 09:30",
-    speaker: "John Doe",
-  },
-  {
-    title: "Intro into parallel programming with python",
-    duration: "09:00 - 09:30",
-    speaker: "Moses Mark",
-  },
-  {
-    title: "Background of python",
-    duration: "09:00 - 09:30",
-    speaker: "James Karen",
-  },
-];
